requirements.txt
tensorflow>=2.10.0
keras>=2.10.0
opencv-python>=4.6.0
imutils>=0.5.4
numpy>=1.23.0
matplotlib>=3.6.0
scikit-learn>=1.1.0
scipy>=1.9.0
Pillow>=9.2.0
tqdm>=4.64.0
seaborn>=0.12.0
